+++ src/core/runtime.js
@@ -3,7 +3,7 @@
 
   if (global.SLINK_EXTENSION) return;
 
-  const VERSION = '0.18.17';
+  const VERSION = '0.18.18';
   const STORAGE_PREFIX = 'slink.';
 
   const runtime = {